# Conflicts:
#	src/main/java/es/fiturjc/restcontroller/AdminRestController.java
#	src/main/java/es/fiturjc/service/UserService.java

+++ src/main/java/es/fiturjc/restcontroller/AdminRestController.java
@@ -24,6 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 import es.fiturjc.component.UserComponent;
 import es.fiturjc.model.Course;
 import es.fiturjc.model.User;
+import es.fiturjc.restcontroller.CourseRestController.CourseDetail;
 import es.fiturjc.service.AdminService;
 import es.fiturjc.service.CourseService;
 import es.fiturjc.service.UserService;
@@ -91,10 +92,14 @@ public class AdminRestController {
     @DeleteMapping(value = "/user/delete/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<?> deleteUser(@PathVariable long id) {
-        if (adminService.deleteUser(id)) {
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	User userLogged = userService.findOne(userComponent.getLoggedUser().getId());
+    	if(userLogged.isAdmin()== true) {
+	        if (adminService.deleteUser(id)) {
+	            return new ResponseEntity<>(HttpStatus.OK);
+	        }
+	        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	}
+    	return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     /**
@@ -123,10 +128,10 @@ public class AdminRestController {
     }
 
 
-    //****************** COURSES **************
+    /****************** COURSES **************/
 
     /**
-     * MIRAR PORQUE NO VA
+     *
      *
      * @return
      */
@@ -134,13 +139,41 @@ public class AdminRestController {
     @RequestMapping(value = "/courses", method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<List<Course>> getCourses() {
+    	User userLogged = userService.findOne(userComponent.getLoggedUser().getId());
         List<Course> courses = courseService.getAllCourses();
-        if (courses != null) {
-            return new ResponseEntity<>(courses, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+    	if(userLogged.isAdmin()) {
+	        if (courses != null) {
+	            return new ResponseEntity<>(courses, HttpStatus.OK);
+	        } else {
+	            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+	        }
+    	} else {
+    		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+    	}
     }
+
+    /**
+	 * Get 1 course
+	 * @param id
+	 * @return
+	 */
+
+	@RequestMapping(value = "/course/{id}", method = RequestMethod.GET)
+	@ResponseStatus(HttpStatus.OK)
+	@JsonView(CourseDetail.class)
+	public ResponseEntity<Course> getCourseId(@PathVariable long id) {
+		User userLogged = userService.findOne(userComponent.getLoggedUser().getId());
+		Course course = courseService.findCourse(id);
+		if(userLogged.isAdmin()) {
+			if (course != null) {
+				return new ResponseEntity<>(course, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+		} else {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+	}
 
     /**
      * @param course
@@ -153,6 +186,7 @@ public class AdminRestController {
     @ResponseStatus(HttpStatus.CREATED)
 
     public ResponseEntity<Course> addCourse(@RequestBody Course course) {
+
         if (userComponent.isLoggedUser()) {
             courseService.save(course);
             return new ResponseEntity<>(course, HttpStatus.OK);
@@ -171,10 +205,15 @@ public class AdminRestController {
     @DeleteMapping(value = "/course/delete/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<?> deleteCourse(@PathVariable long id) {
-        if (courseService.deleteCourse(id)) {
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	User userLogged = userService.findOne(userComponent.getLoggedUser().getId());
+    	if(userLogged.isAdmin()) {
+	        if (courseService.deleteCourse(id)) {
+	            return new ResponseEntity<>(HttpStatus.OK);
+	        }
+	        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	}else {
+    		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+    	}
     }
 
     //to do: solve the way to modify schedules
